chore: update configuration files
Update ESLint, vitest, and prettier configuration files to meet the
latest standards.

Add `engines` field in package.json and let node version to be at least
v20.18.2. This is the node version of the GitHub Action runner as of
this commit (2025/03/27; see
https://github.com/actions/runner/releases/tag/v2.323.0). Setting this
field is helpful to let ESLint to find misuses of unsupported APIs.